add: projectile abilty

+++ src/survivors/Assets/Code/Gameplay/Features/Armaments/ArmamentsComponents.cs
@@ -5,6 +5,6 @@ namespace Code.Gameplay.Features.Armaments
     public class ArmamentsComponents
     {
         [Game] public class Armament : IComponent {  }
-        [Game] public class TargetLimit : IComponent { public float Value; }
+        [Game] public class TargetLimit : IComponent { public int Value; }
     }
 }
